requirements.txt
pyautogen==0.2.27
ag2[gemini]
google-genai
vertexai
jsonschema